pyproject.toml
[project]
name = "storyagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-deepseek>=0.1.3",
    "langchain[deepseek]>=0.3.25",
    "langgraph>=0.4.8",
    "openai>=1.86.0",
    "pytest>=8.4.0",
    "python-dotenv>=1.1.0",
]

[tool.pytest.ini_options]
pythonpath = [
    "."
]
testpaths = [
    "StoryAgent"
]
python_files = ["test_*.py"]
 